# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Internet Weather API",
    "description": "Internet Weather observes the public internet, detects meaningful change in the\nAI-engineering ecosystem, and explains it with traceable evidence.\n\nThe frontend consumes only this API. Provider APIs are never exposed directly.\n\nEvery data response carries a `freshness` object stating when the signals were\ncomputed and how many days of observation support them. Findings carry an\n`epistemic_status` of `observation`, `inference`, `hypothesis` or `unknown`, so\nspeculation is never presented as fact.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/research": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Queue an investigation",
        "description": "Returns immediately with a job. The frontend never blocks on research.",
        "operationId": "create_api_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchJob"
                }
              }
            }
          },
          "501": {
            "description": "The research engine is not implemented in this slice. The contract is final; the implementation follows the first vertical slice."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/{research_id}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Poll an investigation",
        "operationId": "read_api_research__research_id__get",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResult"
                }
              }
            }
          },
          "501": {
            "description": "The research engine is not implemented in this slice. The contract is final; the implementation follows the first vertical slice."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/{research_id}/chat": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Ask the Research Copilot",
        "description": "Contextual to one research topic, grounded in Internet Weather data.\n\nNot a general-purpose chatbot: an answer it cannot ground in stored evidence\nis an answer it must decline to give.",
        "operationId": "chat_api_research__research_id__chat_post",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "title": "Research Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatReply"
                }
              }
            }
          },
          "501": {
            "description": "The research engine is not implemented in this slice. The contract is final; the implementation follows the first vertical slice."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatReply": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "epistemic_status": {
            "$ref": "#/components/schemas/EpistemicStatus"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLink"
            },
            "type": "array",
            "title": "Evidence"
          },
          "data_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Used"
          }
        },
        "type": "object",
        "required": [
          "reply",
          "epistemic_status"
        ],
        "title": "ChatReply",
        "description": "A Copilot answer. Grounded in Internet Weather data or it does not ship."
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
...
      "EpistemicStatus": {
        "type": "string",
        "enum": [
          "observation",
          "inference",
          "hypothesis",
          "unknown"
        ],
        "title": "EpistemicStatus",
        "description": "Required label on every generated claim.\n\nEnforced at the contract level so speculation can never be presented as\nfact. See CLAUDE.md, \"LLM Philosophy\"."
      },
...
      "EvidenceLink": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "The measured value that triggered this, when applicable."
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "EvidenceLink",
        "description": "A traceable source for a claim.\n\nPresent on every event so \"show me the evidence\" is answerable without a\nround trip through the research engine."
      },
      "Finding": {
        "properties": {
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "epistemic_status": {
            "$ref": "#/components/schemas/EpistemicStatus"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLink"
            },
            "type": "array",
            "title": "Evidence"
          },
          "counter_evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLink"
            },
            "type": "array",
            "title": "Counter Evidence"
          }
        },
        "type": "object",
        "required": [
          "statement",
          "epistemic_status",
          "confidence"
        ],
        "title": "Finding",
        "description": "One claim, with its epistemic standing attached.\n\n``epistemic_status`` is required, not optional. That is the mechanism by\nwhich speculation cannot be presented as fact."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchJob": {
        "properties": {
          "research_id": {
            "type": "string",
            "title": "Research Id"
          },
          "status": {
            "$ref": "#/components/schemas/ResearchStatus"
          },
          "technology_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Slug"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "progress_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Note"
          }
        },
        "type": "object",
        "required": [
          "research_id",
          "status"
        ],
        "title": "ResearchJob"
      },
      "ResearchRequest": {
        "properties": {
          "technology_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Slug"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Question",
            "description": "Optional focus, e.g. 'Why is MCP accelerating?'"
          }
        },
        "type": "object",
        "title": "ResearchRequest"
      },
      "ResearchResult": {
        "properties": {
          "research_id": {
            "type": "string",
            "title": "Research Id"
          },
          "status": {
            "$ref": "#/components/schemas/ResearchStatus"
          },
          "technology_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Slug"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "progress_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Note"
          },
          "executive_finding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Finding"
              },
              {
                "type": "null"
              }
            ]
          },
          "supporting_findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Supporting Findings"
          },
          "unknowns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknowns",
            "description": "What the evidence could not settle."
          },
          "watch_next": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Watch Next"
          }
        },
        "type": "object",
        "required": [
          "research_id",
          "status"
        ],
        "title": "ResearchResult"
      },
      "ResearchStatus": {
        "type": "string",
        "enum": [
          "queued",
          "collecting",
          "analyzing",
          "synthesizing",
          "completed",
          "failed"
        ],
        "title": "ResearchStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}